Move v1 files to specific folde
