+++ src/em_backend/v1/README.md
@@ -1,0 +1,4 @@
+# Notice
+
+> Code in this folder comes from the v1 of ElectOMate.
+> There is not guarantee of it working
